+++ test/createVirtualAudioGraph.js
@@ -1,22 +1,22 @@
 /* global AudioContext */
 const test = require('tape')
 require('./WebAudioTestAPISetup')
+const V = require('..')
 
-const createVirtualAudioGraph = require('..')
 const audioContext = new AudioContext()
 
 test('createVirtualAudioGraph - optionally takes audioContext property', t => {
-  t.true(createVirtualAudioGraph({audioContext}).audioContext === audioContext)
-  t.false(createVirtualAudioGraph().audioContext === audioContext)
-  t.true(createVirtualAudioGraph().audioContext instanceof AudioContext)
+  t.true(V.default({audioContext}).audioContext === audioContext)
+  t.false(V.default().audioContext === audioContext)
+  t.true(V.default().audioContext instanceof AudioContext)
   t.end()
 })
 
 test('createVirtualAudioGraph - optionally takes output parameter', t => {
   const gain = audioContext.createGain()
 
-  createVirtualAudioGraph({audioContext, output: gain}).update({
-    0: ['gain', 'output', {gain: 0.2}],
+  V.default({audioContext, output: gain}).update({
+    0: V.gain('output', {gain: 0.2}),
   })
 
   t.deepEqual(gain.toJSON(), {
@@ -30,8 +30,8 @@ test('createVirtualAudioGraph - optionally takes output parameter', t => {
     ],
     name: 'GainNode',
   })
-  createVirtualAudioGraph({audioContext}).update({
-    0: ['gain', 'output', {gain: 0.2}],
+  V.default({audioContext}).update({
+    0: V.gain('output', {gain: 0.2}),
   })
   t.deepEqual(audioContext.toJSON(), {
     inputs: [